Enumeration can be declared before defined.
An unscoped enum can be declared before defined, but the type
must be specified when declared, because the compiler need to
know the size it will take to optimize, for example, it needs
to know if a char is enough or an integer is needed.

An scoped enum can be declared before defined, and a type can
be skipped for it.
If the type is not specified, then the default type will be
int.